# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MediSalud Appointment API
  description: |
    API REST para el sistema de agendamiento de citas medicas de la clinica MediSalud.
    Permite gestionar medicos, pacientes, citas, disponibilidad y festivos de Colombia.
  version: 1.0.0
  contact:
    name: MediSalud Dev Team
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/pacientes:
    post:
      operationId: createPaciente
      tags:
        - Pacientes
      summary: Registrar un nuevo paciente
      description: Crea un nuevo paciente en el sistema con los datos proporcionados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PacienteRequest'
            examples:
              default:
                value:
                  nombre_completo: "Juan Perez Garcia"
                  documento_identidad: "1234567890"
                  telefono: "3001234567"
                  email: "juan.perez@email.com"
                  birth_date: "1990-05-15"
      responses:
        '201':
          description: Paciente creado exitosamente
          headers:
            Location:
              description: URL del recurso creado
              schema:
                type: string
                examples:
                  - /api/v1/pacientes/550e8400-e29b-41d4-a716-446655440000
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PacienteResponse'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Documento de identidad duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      operationId: listPacientes
      tags:
        - Pacientes
      summary: Listar todos los pacientes
      description: Retorna la lista completa de pacientes registrados.
      responses:
        '200':
          description: Lista de pacientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PacienteResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/pacientes/{pacienteId}:
    get:
      operationId: getPaciente
      tags:
        - Pacientes
      summary: Consultar paciente individual
      description: Retorna los datos de un paciente especifico por su ID.
      parameters:
        - $ref: '#/components/parameters/PacienteIdPath'
      responses:
        '200':
          description: Datos del paciente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PacienteResponse'
        '400':
          description: ID invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Paciente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: updatePaciente
      tags:
        - Pacientes
      summary: Actualizar datos de paciente
      description: Actualiza los datos de un paciente existente.
      parameters:
        - $ref: '#/components/parameters/PacienteIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PacienteUpdateRequest'
            examples:
              default:
                value:
                  nombre_completo: "Juan Perez Garcia"
                  documento_identidad: "1234567890"
                  telefono: "3001234567"
                  email: "juan.perez.nuevo@email.com"
                  birth_date: "1990-05-15"
      responses:
        '200':
          description: Paciente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PacienteResponse'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Paciente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Documento de identidad duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    PacienteIdPath:
      name: pacienteId
      in: path
      required: true
      description: ID unico del paciente
      schema:
        type: string
        format: uuid
...
  schemas:
...
    PacienteRequest:
      type: object
      required:
        - nombre_completo
        - documento_identidad
        - telefono
        - email
      properties:
        nombre_completo:
          type: string
          minLength: 3
          maxLength: 100
          description: Nombre completo del paciente
          examples:
            - "Juan Perez Garcia"
        documento_identidad:
          type: string
          minLength: 7
          maxLength: 20
          description: Documento de identidad (unico en el sistema)
          examples:
            - "1234567890"
        telefono:
          type: string
          minLength: 7
          maxLength: 20
          description: Telefono de contacto
          examples:
            - "3001234567"
        email:
          type: string
          format: email
          maxLength: 255
          description: Correo electronico
          examples:
            - "juan.perez@email.com"
        birth_date:
          type: [string, 'null']
          format: date
          description: Fecha de nacimiento (ISO 8601). No puede ser futura. Si no se provee, se asume edad 0.
          examples:
            - "1990-05-15"
    PacienteUpdateRequest:
      type: object
      required:
        - nombre_completo
        - documento_identidad
        - telefono
        - email
      properties:
        nombre_completo:
          type: string
          minLength: 3
          maxLength: 100
          description: Nombre completo del paciente
          examples:
            - "Juan Perez Garcia"
        documento_identidad:
          type: string
          minLength: 7
          maxLength: 20
          description: Documento de identidad
          examples:
            - "1234567890"
        telefono:
          type: string
          minLength: 7
          maxLength: 20
          description: Telefono de contacto
          examples:
            - "3001234567"
        email:
          type: string
          format: email
          maxLength: 255
          description: Correo electronico
          examples:
            - "juan.perez.nuevo@email.com"
        birth_date:
          type: [string, 'null']
          format: date
          description: Fecha de nacimiento
          examples:
            - "1990-05-15"
    PacienteResponse:
      type: object
      required:
        - id
        - nombre_completo
        - documento_identidad
        - telefono
        - email
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: ID unico del paciente
          examples:
            - "550e8400-e29b-41d4-a716-446655440001"
        nombre_completo:
          type: string
          description: Nombre completo
          examples:
            - "Juan Perez Garcia"
        documento_identidad:
          type: string
          description: Documento de identidad
          examples:
            - "1234567890"
        telefono:
          type: string
          description: Telefono de contacto
          examples:
            - "3001234567"
        email:
          type: string
          description: Correo electronico
          examples:
            - "juan.perez@email.com"
        birth_date:
          type: [string, 'null']
          format: date
          description: Fecha de nacimiento
          examples:
            - "1990-05-15"
        created_at:
          type: string
          format: date-time
          description: Fecha de creacion
          examples:
            - "2026-07-17T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Fecha de ultima actualizacion
          examples:
            - "2026-07-17T10:00:00Z"

    # ==================== CITAS ====================
...
    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - code
        - message
        - path
        - trace_id
        - details
      properties:
        timestamp:
          type: string
          format: date-time
          description: Instante UTC del error
          examples:
            - "2026-07-17T10:30:00Z"
        status:
          type: integer
          description: Codigo HTTP de estado
          examples:
            - 400
        error:
          type: string
          description: Nombre estandar del status HTTP
          examples:
            - "BAD_REQUEST"
        code:
          type: string
          description: Codigo interno estable del error
          examples:
            - "VALIDATION_ERROR"
        message:
          type: string
          description: Mensaje descriptivo del error
          examples:
            - "The request contains invalid fields."
        path:
          type: string
          description: Ruta del endpoint que genero el error
          examples:
            - "/api/v1/medicos"
        trace_id:
          type: string
          description: Identificador unico de la traza
          examples:
            - "4f0f6d2c6b1d4c8a"
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiErrorDetail'
          description: Lista de detalles adicionales
    ApiErrorDetail:
      type: object
      properties:
        field:
          type: [string, 'null']
          description: Campo con error (si aplica)
          examples:
            - "nombre_completo"
        code:
          type: string
          description: Codigo del error de campo
          examples:
            - "FIELD_INVALID"
        message:
          type: string
          description: Mensaje del error de campo
          examples:
            - "must be between 3 and 100 characters"
        rejected_value:
          type: [string, 'null']
          description: Valor rechazado (si no expone datos sensibles)
          examples:
            - "AB"
  responses:
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            default:
              value:
                timestamp: "2026-07-17T10:30:00Z"
                status: 500
                error: "INTERNAL_SERVER_ERROR"
                code: "INTERNAL_ERROR"
                message: "An unexpected error occurred."
                path: "/api/v1/medicos"
                trace_id: "4f0f6d2c6b1d4c8a"
                details: []